+++ Operating Systems project 2/Logic/FileReader.Logic.cs
@@ -7,8 +7,8 @@ namespace Operating_Systems_Project
     {
         private static string ReadTextFile(string path)
         {
-            using (FileStream fileS = new FileStream(path, FileMode.Open, FileAccess.Read))
-            using (BufferedStream buffer = new BufferedStream(fileS))
+            using (FileStream fs = new FileStream(path, FileMode.Open, FileAccess.Read))
+            using (BufferedStream buffer = new BufferedStream(fs))
             using (StreamReader reader = new StreamReader(buffer))
                 return reader.ReadToEnd();
         }
@@ -18,8 +18,23 @@ namespace Operating_Systems_Project
             return File.ReadAllText(path);
         }
 
+        private static int ReadBinaryFileAsInt32(string path) // Returns text
+        {
+            // =========== Important notes ===========
+            // The file must contain at least 4 bytes
+            // The file must have been written as an int, not text
+            // This will not work for text files
+            using (FileStream fs = new FileStream(path, FileMode.Open, FileAccess.Read))
+            using (BinaryReader reader = new BinaryReader(fs))
+                return reader.ReadInt32();
+        }
+
+        // =====================
+        // These Methods are extra ( We didn't take them in the section)
+        // =====================
         private static string ReadBinaryFile(string path) // Returns text
         {
+            // ReadAllText() is a simpler version of this
             byte[] bytes;
 
             using (BinaryReader reader = new BinaryReader(File.Open(path, FileMode.Open)))